more cleanup to reduce size of single net label placement solver
