+++ lib/solvers/NetLabelPlacementSolver/SingleNetLabelPlacementSolver/SingleNetLabelPlacementSolver.ts
@@ -9,18 +9,16 @@ import type { MspConnectionPairId } from "lib/solvers/MspConnectionPairSolver/Ms
 import type { FacingDirection } from "lib/utils/dir"
 import type { GraphicsObject } from "graphics-debug"
 import { ChipObstacleSpatialIndex } from "lib/data-structures/ChipObstacleSpatialIndex"
-import { visualizeInputProblem } from "lib/solvers/SchematicTracePipelineSolver/visualizeInputProblem"
-import { getColorFromString } from "lib/utils/getColorFromString"
 import {
   getDimsForOrientation,
   getCenterFromAnchor,
   getRectBounds,
-  NET_LABEL_HORIZONTAL_WIDTH,
-  NET_LABEL_HORIZONTAL_HEIGHT,
 } from "./geometry"
 import { rectIntersectsAnyTrace } from "./collisions"
 import { chooseHostTraceForGroup } from "./host"
 import { anchorsForSegment } from "./anchors"
+import { solveNetLabelPlacementForPortOnlyPin } from "./solvePortOnlyPin"
+import { visualizeSingleNetLabelPlacementSolver } from "./SingleNetLabelPlacementSolver_visualize"
 
 export {
   NET_LABEL_HORIZONTAL_WIDTH,
@@ -97,119 +95,22 @@ export class SingleNetLabelPlacementSolver extends BaseSolver {
 
     // Handle port-only island (no traces) by placing a label at the port
     if (this.overlappingSameNetTraceGroup.portOnlyPinId) {
-      const pinId = this.overlappingSameNetTraceGroup.portOnlyPinId
-      // Find pin coordinates
-      let pin: { x: number; y: number } | null = null
-      for (const chip of this.inputProblem.chips) {
-        const p = chip.pins.find((pp) => pp.pinId === pinId)
-        if (p) {
-          pin = { x: p.x, y: p.y }
-          break
-        }
-      }
-      if (!pin) {
-        this.failed = true
-        this.error = `Port-only pin not found: ${pinId}`
-        return
-      }
-
-      const orientations =
-        this.availableOrientations.length > 0
-          ? this.availableOrientations
-          : (["x+", "x-", "y+", "y-"] as FacingDirection[])
-
-      const anchor = { x: pin.x, y: pin.y }
-      const outwardOf = (o: FacingDirection) =>
-        o === "x+"
-          ? { x: 1, y: 0 }
-          : o === "x-"
-            ? { x: -1, y: 0 }
-            : o === "y+"
-              ? { x: 0, y: 1 }
-              : { x: 0, y: -1 }
-
-      for (const orientation of orientations) {
-        const { width, height } = getDimsForOrientation(orientation)
-        // Place label fully outside the chip: shift center slightly outward
-        const baseCenter = getCenterFromAnchor(
-          anchor,
-          orientation,
-          width,
-          height,
-        )
-        const outward = outwardOf(orientation)
-        const offset = 1e-3
-        const center = {
-          x: baseCenter.x + outward.x * offset,
-          y: baseCenter.y + outward.y * offset,
-        }
-        const bounds = getRectBounds(center, width, height)
-
-        // Chip collision check
-        const chips = this.chipObstacleSpatialIndex.getChipsInBounds(bounds)
-        if (chips.length > 0) {
-          this.testedCandidates.push({
-            center,
-            width,
-            height,
-            bounds,
-            anchor,
-            orientation,
-            status: "chip-collision",
-            hostSegIndex: -1,
-          })
-          continue
-        }
-
-        // Trace collision check
-        if (
-          rectIntersectsAnyTrace(
-            bounds,
-            this.inputTraceMap,
-            "" as MspConnectionPairId,
-            -1,
-          )
-        ) {
-          this.testedCandidates.push({
-            center,
-            width,
-            height,
-            bounds,
-            anchor,
-            orientation,
-            status: "trace-collision",
-            hostSegIndex: -1,
-          })
-          continue
-        }
-
-        // Found a valid placement
-        this.testedCandidates.push({
-          center,
-          width,
-          height,
-          bounds,
-          anchor,
-          orientation,
-          status: "ok",
-          hostSegIndex: -1,
-        })
-
-        this.netLabelPlacement = {
-          globalConnNetId: this.overlappingSameNetTraceGroup.globalConnNetId,
-          dcConnNetId: undefined,
-          orientation,
-          anchorPoint: anchor,
-          width,
-          height,
-          center,
-        }
+      const res = solveNetLabelPlacementForPortOnlyPin({
+        inputProblem: this.inputProblem,
+        inputTraceMap: this.inputTraceMap,
+        chipObstacleSpatialIndex: this.chipObstacleSpatialIndex,
+        overlappingSameNetTraceGroup: this.overlappingSameNetTraceGroup,
+        availableOrientations: this.availableOrientations,
+      })
+      this.testedCandidates.push(...res.testedCandidates)
+      if (res.placement) {
+        this.netLabelPlacement = res.placement
         this.solved = true
         return
       }
-
       this.failed = true
-      this.error = "Could not place net label at port without collisions"
+      this.error =
+        res.error ?? "Could not place net label at port without collisions"
       return
     }
 
@@ -352,84 +253,6 @@ export class SingleNetLabelPlacementSolver extends BaseSolver {
   }
 
   override visualize(): GraphicsObject {
-    const graphics = visualizeInputProblem(this.inputProblem)
-
-    // Visualize the entire trace group for this net id
-    const groupId = this.overlappingSameNetTraceGroup.globalConnNetId
-    const host = chooseHostTraceForGroup({
-      inputProblem: this.inputProblem,
-      inputTraceMap: this.inputTraceMap,
-      globalConnNetId: groupId,
-      fallbackTrace: this.overlappingSameNetTraceGroup.overlappingTraces,
-    })
-    const groupStroke = getColorFromString(groupId, 0.9)
-    const groupFill = getColorFromString(groupId, 0.5)
-
-    for (const trace of Object.values(this.inputTraceMap)) {
-      if (trace.globalConnNetId !== groupId) continue
-      const isHost = host ? trace.mspPairId === host.mspPairId : false
-      graphics.lines!.push({
-        points: trace.tracePath,
-        strokeColor: isHost ? groupStroke : groupFill,
-        strokeWidth: isHost ? 0.006 : 0.003,
-        strokeDash: isHost ? undefined : "4 2",
-      } as any)
-    }
-
-    // Visualize all tested candidate rectangles with reason coloring
-    for (const c of this.testedCandidates) {
-      const fill =
-        c.status === "ok"
-          ? "rgba(0, 180, 0, 0.25)"
-          : c.status === "chip-collision"
-            ? "rgba(220, 0, 0, 0.25)"
-            : c.status === "trace-collision"
-              ? "rgba(220, 140, 0, 0.25)"
-              : "rgba(120, 120, 120, 0.15)"
-      const stroke =
-        c.status === "ok"
-          ? "green"
-          : c.status === "chip-collision"
-            ? "red"
-            : c.status === "trace-collision"
-              ? "orange"
-              : "gray"
-
-      graphics.rects!.push({
-        center: {
-          x: (c.bounds.minX + c.bounds.maxX) / 2,
-          y: (c.bounds.minY + c.bounds.maxY) / 2,
-        },
-        width: c.width,
-        height: c.height,
-        fill,
-        strokeColor: stroke,
-      } as any)
-
-      graphics.points!.push({
-        x: c.anchor.x,
-        y: c.anchor.y,
-        color: stroke,
-      } as any)
-    }
-
-    // Visualize the final accepted label (if any)
-    if (this.netLabelPlacement) {
-      const p = this.netLabelPlacement
-      graphics.rects!.push({
-        center: p.center,
-        width: p.width,
-        height: p.height,
-        fill: "rgba(0, 128, 255, 0.35)",
-        strokeColor: "blue",
-      } as any)
-      graphics.points!.push({
-        x: p.anchorPoint.x,
-        y: p.anchorPoint.y,
-        color: "blue",
-      } as any)
-    }
-
-    return graphics
+    return visualizeSingleNetLabelPlacementSolver(this)
   }
 }
